Support dtype=np.float32 even when jax_enable_x64 is set True
To enable double precision float in JAX, use:

    from jax.config import config
    config.update("jax_enable_x64", True)

+++ mod/fadge/geode.py
@@ -45,7 +45,7 @@ def JA(metric):
         a  = (-       dot(ig, dot(dot(dg, v), v))
               + 0.5 * dot(ig, dot(v, dot(v, dg))))
 
-        return np.array([v, a])
+        return np.array([v, a], dtype=state.dtype)
 
     return rhs
